# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  api.MapPoint:
    properties:
      elevation:
        type: number
      lat:
        type: number
      lng:
        type: number
      total_distance:
        type: number
    type: object
...
  api.PaginatedResponse-api_RouteSegmentResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      page:
        type: integer
      per_page:
        type: integer
      results:
        items:
          $ref: '#/definitions/api.RouteSegmentResponse'
        type: array
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  api.Response-any:
    properties:
      errors:
        items:
          type: string
        type: array
      results: {}
    type: object
...
  api.Response-api_RouteSegmentDetailResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        $ref: '#/definitions/api.RouteSegmentDetailResponse'
    type: object
  api.Response-api_RouteSegmentsDetailResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        items:
          $ref: '#/definitions/api.RouteSegmentResponse'
        type: array
    type: object
...
  api.Response-map_string_string:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        $ref: '#/definitions/map_string_string'
    type: object
  api.RouteSegmentDetailResponse:
    properties:
      address_string:
        type: string
      bidirectional:
        type: boolean
      center:
        properties:
          lat:
            type: number
          lng:
            type: number
        type: object
      circular:
        type: boolean
      created_at:
        type: string
      filename:
        type: string
      id:
        type: integer
      match_count:
        type: integer
      matches:
        items:
          $ref: '#/definitions/api.RouteSegmentMatch'
        type: array
      max_elevation:
        type: number
      min_elevation:
        type: number
      name:
        type: string
      notes:
        type: string
      points:
        items:
          $ref: '#/definitions/api.MapPoint'
        type: array
      total_distance:
        type: number
      total_down:
        type: number
      total_up:
        type: number
      updated_at:
        type: string
    type: object
  api.RouteSegmentMatch:
    properties:
      average_speed:
        type: number
      distance:
        type: number
      duration:
        type: integer
      user_id:
        type: integer
      user_name:
        type: string
      workout_id:
        type: integer
      workout_name:
        type: string
    type: object
...
  api.RouteSegmentResponse:
    properties:
      bidirectional:
        type: boolean
      circular:
        type: boolean
      created_at:
        type: string
      filename:
        type: string
      id:
        type: integer
      match_count:
        type: integer
      max_elevation:
        type: number
      min_elevation:
        type: number
      name:
        type: string
      notes:
        type: string
      total_distance:
        type: number
      total_down:
        type: number
      total_up:
        type: number
      updated_at:
        type: string
    type: object
...
  map_string_string:
    additionalProperties:
      type: string
    type: object
info:
  contact: {}
  description: Workout Tracker HTTP API (v2).
  title: Workout Tracker API
  version: "2.0"
paths:
...
  /route-segments:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaginatedResponse-api_RouteSegmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: List route segments
      tags:
      - route-segments
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: GPX file
        in: formData
        name: file
        required: true
        type: file
      - description: Notes
        in: formData
        name: notes
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Response-api_RouteSegmentsDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Create route segment
      tags:
      - route-segments
  /route-segments/{id}:
    delete:
      parameters:
      - description: Route segment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-map_string_string'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Delete route segment
      tags:
      - route-segments
    get:
      parameters:
      - description: Route segment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_RouteSegmentDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Get route segment
      tags:
      - route-segments
    put:
      consumes:
      - application/json
      parameters:
      - description: Route segment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_RouteSegmentDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Update route segment
      tags:
      - route-segments
  /route-segments/{id}/download:
    get:
      parameters:
      - description: Route segment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: binary GPX content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Download route segment file
      tags:
      - route-segments
  /route-segments/{id}/matches:
    post:
      parameters:
      - description: Route segment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-map_string_string'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Find matching workouts
      tags:
      - route-segments
  /route-segments/{id}/refresh:
    post:
      parameters:
      - description: Route segment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-map_string_string'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Refresh route segment
      tags:
      - route-segments
...
  /workouts/{id}/route-segment:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workout ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Response-api_RouteSegmentDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Create route segment from workout
      tags:
      - route-segments
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  ApiKeyQuery:
    in: query
    name: api-key
    type: apiKey
  CookieAuth:
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"